% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvd_concept_set}
\alias{cvd_concept_set}
\title{CVD Sample Concept Set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 43 rows and 4 columns.
}
\usage{
cvd_concept_set
}
\description{
A sample version of the table structure expected for the \code{concept_set}
parameter in the \code{cvd_process} function. The user should recreate
this file and include their own clinical
}
\details{
@format ## cvd_concept_set
A data frame with 4 columns:
\describe{
\item{concept_id}{The OMOP concept_id; if the PCORnet CDM is being used, default this column to a random integer like the row number}
\item{concept_name}{(optional)The string name of the concept}
\item{concept_code}{The original code associated with the concept_id}
\item{vocabulary_id}{The vocabulary associated with the concept; if the PCORnet CDM is being used, ensure that the values of this field match the vocabulary abbreviations used in the CDM itself}
}
}
\keyword{datasets}
