% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvd_domain_file}
\alias{cvd_domain_file}
\title{CVD Sample Domain File}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.
}
\usage{
cvd_domain_file
}
\description{
A sample version of the table structure expected for the \code{domain_tbl}
parameter in the \code{cvd_process} function. The user should recreate
this table and include their own domain definitions.
}
\details{
#' @format ## cvd_domain_file
\describe{
A data frame with 4 columns
\item{domain}{The name of the CDM table associated with the domain of interest}
\item{concept_field}{The name of the column in the domain table that contains the concepts of interest listed in the concept_set file.}
\item{vs_field}{The name of the column in the domain table for which to show a distribution of categorical values.}
\item{date_field}{The name of the column in the domain table that contains dates to be used for time-based filtering.}
}
}
\keyword{datasets}
