% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output_subfuncs.R
\name{cvd_ms_anom_cs}
\alias{cvd_ms_anom_cs}
\title{Multi Site, Anomaly, Cross-Sectional Output for Categorical Variable Distribution Module}
\usage{
cvd_ms_anom_cs(process_output)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}
}
\value{
A dot plot, returned as a ggplot object, with site on x axis, categorical valueset item on y axis, with dot size representing mean proportion for the given valueset item across all sites, color representing proportion of given site's values for the given valueset item. Anomalies are distinguished with star shapes as opposed to the normal circles.
}
\description{
Multi Site, Anomaly, Cross-Sectional Output for Categorical Variable Distribution Module
}
