% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output_subfuncs.R
\name{cvd_ms_anom_la}
\alias{cvd_ms_anom_la}
\title{Multi Site, Anomaly, Longitudinal Output for Categorical Variable Distribution Module}
\usage{
cvd_ms_anom_la(process_output, filt)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}

\item{filt}{\emph{numeric/string or vector} | The specific code(s) that should be the focus of the analysis}
}
\value{
three graphs:
\enumerate{
\item Loess smoothed line graph that shows the proportion of a code across time
with the Euclidean Distance associated with each line
\item same as (1) but displaying the raw, unsmoothed proportion
\item a radial bar graph displaying the Euclidean Distance value for each
site, where the color is the average proportion across time
}
}
\description{
Multi Site, Anomaly, Longitudinal Output for Categorical Variable Distribution Module
}
