% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output_subfuncs.R
\name{cvd_ms_exp_cs}
\alias{cvd_ms_exp_cs}
\title{Multi Site, Exploratory, Cross-Sectional Output for Categorical Variable Distribution Module}
\usage{
cvd_ms_exp_cs(process_output)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}
}
\value{
A heatmap, returned as a ggplot object, with site on x axis, valueset item on y axis, with proportion of the site's total for each valueset item displayed as tile color and a label on the tile
}
\description{
Multi Site, Exploratory, Cross-Sectional Output for Categorical Variable Distribution Module
}
