% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output_subfuncs.R
\name{cvd_ms_exp_la}
\alias{cvd_ms_exp_la}
\title{Multi Site, Exploratory, Longitudinal Output for Categorical Variable Distribution Module}
\usage{
cvd_ms_exp_la(process_output)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}
}
\value{
A line plot, returned as a ggplot object, faceted by categorical valueset item, with lines and line color representing site.
}
\description{
Multi Site, Exploratory, Longitudinal Output for Categorical Variable Distribution Module
}
