% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output.R
\name{cvd_output}
\alias{cvd_output}
\title{Categorical Variable Distribution - Output Generation}
\usage{
cvd_output(process_output, filter_concept)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}

\item{filter_concept}{\emph{numeric/string or vector} | The specific code(s) that should be the focus of the analysis; used for \code{ss_anom_la} and \code{ms_anom_la}}
}
\value{
a graph or table to visualize the results from
}
\description{
Using the tabular output generated by \code{cvd_process}, this function will build a visualization of the results
Each function configuration will output a bespoke ggplot object.
Plot theming can be adjusted by the user after the graph has bee output using \code{+ theme()}.
Plots can also be made interactive by using \code{make_interactive_squba()} on the object output from \code{cvd_output}.
}
