% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_process.R
\name{cvd_process}
\alias{cvd_process}
\title{Categorical Variable Distribution Module Processing}
\usage{
cvd_process(
  cohort,
  domain_tbl = NULL,
  concept_set = NULL,
  omop_or_pcornet,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  p_value = 0.9,
  time = TRUE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year",
  vocab_tbl = NULL
)
}
\arguments{
\item{cohort}{\emph{tabular input} | A dataframe with the cohort of patients for your study. Should include the columns:
\itemize{
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character}
\item \code{start_date} | \emph{date}
\item \code{end_date} | \emph{date}
\item \code{site} | \emph{character}
}}

\item{domain_tbl}{\emph{tabular input} | A table with domain definitions. An example file is provided as \code{categoricalvariabledistribution::cvd_domain_file}. Should include the columns:
\itemize{
\item \code{domain} | The name of the CDM table associated with the domain of interest
\item \code{concept_field} | The name of the column in the domain table that contains the concepts of interest listed in the concept_set file
\item \code{vs_field} | The name of the column in the domain table for which to show a distribution of categorical values.
\code{date_field} | The name of the column in the domain table that contains dates to be used for time-based filtering
}}

\item{concept_set}{\emph{tabular input} | An annotated concept set with the following columns:
\itemize{
\item \code{concept_id} | \emph{integer} |  required for OMOP; the concept_id of interest
\item \code{concept_name} | \emph{character} | optional; the descriptive name of the concept
\item \code{concept_code} | \emph{character} | required for PCORnet; the code of interest
\item \code{vocabulary_id} | \emph{character} | required for PCORnet; the vocabulary of the code - should match what is listed in the domain table's vocabulary_field
}}

\item{omop_or_pcornet}{\emph{string} | Option to run the function using the OMOP or PCORnet CDM as the default CDM
\itemize{
\item \code{omop}: run the \code{\link[=cvd_process_omop]{cvd_process_omop()}} function against an OMOP CDM instance
\item \code{pcornet}: run the \code{\link[=cvd_process_pcornet]{cvd_process_pcornet()}} function against a PCORnet CDM instance
}}

\item{multi_or_single_site}{\emph{string} | Option to run the function on a single vs multiple sites
\itemize{
\item \code{single}: run the function for a single site
\item \code{multi}: run the function for multiple sites
}}

\item{anomaly_or_exploratory}{\emph{string} | Option to conduct an exploratory or anomaly detection analysis.
Exploratory analyses give a high level summary of the data to examine
the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.}

\item{p_value}{\emph{numeric} | the p value to be used as a threshold in the multi-site anomaly detection analysis}

\item{time}{\emph{boolean} | logical to determine whether to output the check across time}

\item{time_span}{\emph{vector - length 2} | when time = TRUE, a vector of two dates for the observation period of the study}

\item{time_period}{\emph{string} | when time = TRUE, this argument defines the distance
between dates within the specified time period.
Defaults to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}

\item{vocab_tbl}{\emph{tabular input} | optional; table containing crosswalk r}
}
\value{
a data frame with summary results that can be used for \code{cvd_output} to generate graphical or tabular output
}
\description{
This is a module that takes a concept set and a field of interest as input
and creates a distribution of the categorical values in the field of interest.
For example, a user can input a drug concept set and get back a distribution
of units associated with the drug, or a user can input a lab test concept set
and get back a distribution of units associated with the lab results.
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'categoricalvariabledistribution'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'cvd_process_test',
                      working_directory = getwd(),
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = 'extdata',
                      cdm_schema = NA)

#' Build mock study cohort
cohort_tbl <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Execute `cvd_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
cvd_process_example <- cvd_process(cohort=cohort_tbl,
                                  domain_tbl=categoricalvariabledistribution::cvd_domain_tbl,
                                  concept_set=categoricalvariabledistribution::cvd_concept_set,
                                  omop_or_pcornet='omop',
                                  multi_or_single_site = 'single',
                                  anomaly_or_exploratory='exploratory',
                                  time = FALSE)

cvd_process_example

#' Execute `cvd_output` function
#' The output was edited for a better indication of what the visualization will
#' look like.
#' The 0s are a limitation of the small sample data set used for this example
cvd_output_example <- cvd_output(process_output = cvd_process_example)

cvd_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(cvd_output_example)
}
