% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_process_pcornet.R
\name{cvd_process_pcornet}
\alias{cvd_process_pcornet}
\title{Categorical Variable Distribution - PCORnet Version}
\usage{
cvd_process_pcornet(
  cohort,
  domain_tbl,
  concept_set,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  p_value = 0.9,
  time = TRUE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year"
)
}
\arguments{
\item{cohort}{\emph{tabular input} | A dataframe with the cohort of patients for your study. Should include the columns:
\itemize{
\item \code{person_id} / \code{patid} | \emph{integer} / \emph{character}
\item \code{start_date} | \emph{date}
\item \code{end_date} | \emph{date}
\item \code{site} | \emph{character}
}}

\item{domain_tbl}{\emph{tabular input} | A table with domain definitions. An example file is provided as \code{categoricalvariabledistribution::cvd_domain_file}. Should include the columns:
\itemize{
\item \code{domain} | The name of the CDM table associated with the domain of interest
\item \code{concept_field} | The name of the column in the domain table that contains the concepts of interest listed in the concept_set file
\item \code{vs_field} | The name of the column in the domain table for which to show a distribution of categorical values.
\code{date_field} | The name of the column in the domain table that contains dates to be used for time-based filtering
}}

\item{concept_set}{\emph{tabular input} | An annotated concept set with the following columns:
\itemize{
\item \code{concept_id} | \emph{integer} |  required for OMOP; the concept_id of interest
\item \code{concept_name} | \emph{character} | optional; the descriptive name of the concept
\item \code{concept_code} | \emph{character} | required for PCORnet; the code of interest
\item \code{vocabulary_id} | \emph{character} | required for PCORnet; the vocabulary of the code - should match what is listed in the domain table's vocabulary_field
}}

\item{multi_or_single_site}{\emph{string} | Option to run the function on a single vs multiple sites
\itemize{
\item \code{single}: run the function for a single site
\item \code{multi}: run the function for multiple sites
}}

\item{anomaly_or_exploratory}{\emph{string} | Option to conduct an exploratory or anomaly detection analysis.
Exploratory analyses give a high level summary of the data to examine
the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.}

\item{p_value}{\emph{numeric} | the p value to be used as a threshold in the multi-site anomaly detection analysis}

\item{time}{\emph{boolean} | logical to determine whether to output the check across time}

\item{time_span}{\emph{vector - length 2} | when time = TRUE, a vector of two dates for the observation period of the study}

\item{time_period}{\emph{string} | when time = TRUE, this argument defines the distance
between dates within the specified time period.
Defaults to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}
}
\value{
a data frame with the distribution from the specified domain/concept which will be further summarized in \code{cvd_process}
}
\description{
Categorical Variable Distribution - PCORnet Version
}
