% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output_subfuncs.R
\name{cvd_ss_anom_la}
\alias{cvd_ss_anom_la}
\title{Single Site, Anomaly, Longitudinal Output for Categorical Variable Distribution Module}
\usage{
cvd_ss_anom_la(process_output, filt)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}
}
\value{
if analysis was executed by year or greater, a P Prime control chart
is returned with outliers marked with orange dots

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if analysis was executed by month or smaller, an STL regression is
    conducted and outliers are marked with red dots. the graphs representing
    the data removed in the regression are also returned
}\if{html}{\out{</div>}}
}
\description{
Single Site, Anomaly, Longitudinal Output for Categorical Variable Distribution Module
}
