% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output_subfuncs.R
\name{cvd_ss_exp_cs}
\alias{cvd_ss_exp_cs}
\title{Single Site, Exploratory, Cross-Sectional Output for Categorical Variable Distribution Module}
\usage{
cvd_ss_exp_cs(process_output)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}
}
\value{
a bar plot as a ggplot object with categorical value set on y axis, proportion on y axis, filled by categorical value
}
\description{
Single Site, Exploratory, Cross-Sectional Output for Categorical Variable Distribution Module
}
