% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output_subfuncs.R
\name{cvd_ss_exp_la}
\alias{cvd_ss_exp_la}
\title{Single Site, Exploratory, Longitudinal Output for Categorical Variable Distribution Module}
\usage{
cvd_ss_exp_la(process_output)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}
}
\value{
A line plot as a ggplot object with time on x axis, proportion of category with value on y axis, colored by categorical value
}
\description{
Single Site, Exploratory, Longitudinal Output for Categorical Variable Distribution Module
}
