% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvd_output.R
\name{cvd_output}
\alias{cvd_output}
\title{Categorical Variable Distribution - Output Generation}
\usage{
cvd_output(process_output, filter_concept, vocab_tbl = NULL)
}
\arguments{
\item{process_output}{\emph{tabular output} | The output from \code{cvd_process}}

\item{filter_concept}{\emph{numeric/string or vector} | The specific code(s) that should be the focus of the analysis; used for \code{ss_anom_la} and \code{ms_anom_la}}

\item{vocab_tbl}{\emph{tabular input} | OPTIONAL: the location of an external vocabulary table containing concept names for
the provided codes. if not NULL, concept names will be available in either a reference
table or in a hover tooltip}
}
\value{
a graph or table to visualize the results from
}
\description{
Using the tabular output generated by \code{cvd_process}, this function will build a visualization of the results
Each function configuration will output a bespoke ggplot object.
Plot theming can be adjusted by the user after the graph has bee output using \code{+ theme()}.
Plots can also be made interactive by using \code{make_interactive_squba()} on the object output from \code{cvd_output}.
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'categoricalvariabledistribution'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'cvd_process_test',
                      working_directory = getwd(),
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = 'extdata',
                      cdm_schema = NA)

#' Build mock study cohort
cohort_tbl <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Execute `cvd_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
cvd_process_example <- cvd_process(cohort=cohort_tbl,
                                  domain_tbl=categoricalvariabledistribution::cvd_domain_tbl,
                                  concept_set=categoricalvariabledistribution::cvd_concept_set,
                                  omop_or_pcornet='omop',
                                  multi_or_single_site = 'single',
                                  anomaly_or_exploratory='exploratory',
                                  time = FALSE)

cvd_process_example

#' Execute `cvd_output` function
#' The output was edited for a better indication of what the visualization will
#' look like.
#' The 0s are a limitation of the small sample data set used for this example
cvd_output_example <- cvd_output(process_output = cvd_process_example)

cvd_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(cvd_output_example)
}
